Cleanup some code
