--- test/support/auth.ex
@@ -1,5 +1,0 @@
-defmodule Auth do
-  use Authsense,
-    repo: Authsense.Test.Repo,
-    model: Authsense.Test.User
-end
